refactor: Pindahkan hook booking ke dalam folder yang lebih terstruktur
- page.tsx: Perbarui import untuk useBooking dari lokasi baru.
- create/page.tsx: Sesuaikan import useAdminBooking ke lokasi baru.
- BookingForm.tsx: Perbarui import useAdminBooking ke lokasi baru.
- BookingFilters.tsx dan BookingFilterUI.tsx: Sesuaikan import BookingFilters ke lokasi baru.
- FieldAvailability.tsx: Perbarui import useBooking ke lokasi baru.
- useDurationCalculator.hook.ts dan useTimeSlot.hook.ts: Sesuaikan import useBooking ke lokasi baru.
- Hapus file useAdminBooking.hook.ts, useBooking.hook.ts, useBookingData.hook.ts, useBookingFilters.hook.ts yang tidak lagi digunakan.
